task3/pyproject.toml
[project]
name = "d2"
version = "0.1.0"
description = ""
authors = [
    {name = "Dion Lind",email = "dion.lind22@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[all] (>=0.117.1,<0.118.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "bcrypt (==3.2.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
